feat: InterviewCategory 생성 코드 작성

+++ til-domain/src/main/java/com/til/domain/interview/dto/InterviewCreateDto.java
@@ -9,7 +9,7 @@ import lombok.Builder;
 
 @Builder
 public record InterviewCreateDto(
-                                 List<Integer> categoryIdList,
+                                 List<Long> categoryIdList,
                                  InterviewStatus status,
                                  String code,
                                  Long userId
